+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## 1.1.5
+
+---
+
 ## 1.1.4
 
 ### Fixes